requirements.txt
aiomysql==0.0.22
aiosqlite==0.17.0
anyio==3.5.0
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
Automat==20.2.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
constantly==15.1.0
decorator==5.1.1
dnspython==2.2.0
email-validator==1.1.3
Flask==2.0.3
flask-swagger-ui==3.36.0
greenlet==1.1.2
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
iniconfig==1.1.1
itsdangerous==2.1.0
Jinja2==3.0.3
keyboard==0.13.5
MarkupSafe==2.1.0
outcome==1.1.0
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pydantic==1.9.0
PyMySQL==0.9.3
pyparsing==3.0.7
pytest==7.0.1
pytest-asyncio==0.18.1
pytest-tornasync==0.6.0.post2
pytest-trio==0.7.0
pytest-twisted==1.13.4
python-dotenv==0.19.2
python-worker==1.8.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.31
tomli==2.0.1
tornado==6.1
trio==0.20.0
Twisted==22.1.0
typing_extensions==4.1.1
Werkzeug==2.0.3
zope.interface==5.4.0
